Don't use f64 in vq codebooks and in residues
It was thought that f64 was needed as the vorbis 32 bit
representation of floats enocdes them differently so that
an exact mapping is only doable by f32 floats. But seems the
f64 floats aren't even really required, as test files appear
to still work. Therefore, using f64 is a waste.

Speedup is minimal though, maybe 1-2%, at best.