More convenient new constructor for OggStreamReader
The goal of OggStreamReader is to be as convenient as possible.
This change makes OggStreamReader easier to use as it creates
the PacketReader for you. This removes the need to import it,
and to add the ogg as a direct dependency to a crate that
uses lewton+ogg.

This is a breaking change, but worth it I think, as the goal
is a convenient OggStreamReader interface, and not stability.