Make blocksize values pub(crate)
Its most likely not very useful for users to find out
the used blocksize values. They might be used to tell
the possible decoded packet sizes in advance,
but they don't represent the final values, as the
final packet size is also affected by windowing,
and in the case of interleaved packets, the number
of channels.