Fix #33 crash
The spec recommends us to clamp
"each element to [0, [range]) after step 1".
However, we clamped to [0, [range]] instead.
This caused the panic when the value got
clamped to 256, and used to derive y values
for the floor that resulted in 256 indices
into the FLOOR1_INVERSE_DB_TABLE array.

Thus, the file is actually valid,
not malformed.

The panic is a bug nontheless. We fix it
by performing the recommended clamping.