Fix single-code-{ordered,sparse}.ogg decoding error
Decoding of the files single-code-ordered.ogg
and single-code-sparse.ogg gave us the error
that current_length was bigger than
codebook_entries. But this was a bug.

Apparently we were comparing current_length
with codebook_entries, but the spec tells us
to use current_entry instead of current_length.

Now, after having changed the behaviour to
follow the spec, we can decode the files
successfully.

The second file can be decoded by libvorbis
as well, and here some discrepancy still seems
to exist (25 packets with mismatches of 26 total).
However, having the files decode successfully
is already progress.