Remove cbindgen again
It seems that we can't use cbindgen because of its
rustc version policy [1] seeming to require always
the latest version. This is incompatible with our
requirements and due to that, we won't be able
to use cbindgen going forward.
Fortunately, the C API that I want lewton to expose
only has to be minimal.

[1]: https://github.com/eqrion/cbindgen/commit/191d8ca91735c30916edfd55f96678f34d2b5953